requirements.txt
paho-mqtt==1.3.1
PyYAML==3.12
click==6.7
dependency_injector==3.11.3
